backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
anthropic==0.43.0
reportlab==4.1.0
python-dotenv==1.0.1
sse-starlette==2.0.0
httpx==0.27.0
pydantic==2.6.1
